make teest mode, and return status for lambda function
