+++ async_destroy_lambda/.envrc
@@ -5,4 +5,5 @@ export GIT_URL='https://github.com/ministryofjustice/opg-lpa.git'
 export REPO_DIR='/tmp/opg-lpa'
 export TF_CONFIG_PATH=terraform/terraform_environment
 export TERRAFORM_VERSION='0.12.5'
+export TEST='True'
